update code 15/5/2021

+++ utils/index.js
@@ -1,24 +1,26 @@
 import AsyncStorage from "@react-native-async-storage/async-storage";
-import { ACCESS_TOKEN, URL_SERVER } from "../constants";
+import UserApi from "../Api/userApi";
+import { STATUS_MESSAGE, URL_SERVER } from "../constants";
 import { SET_ACCESS_TOKEN, SET_IS_LOGIN } from "../redux/actions/Login";
+import store from "../redux/store";
 
 export const isEmpty = (text) => {
   return text === null || text === undefined || text === "";
 };
 
-export const getTokenFromStorage = async () => {
+export const getTokenFromStorage = async (name) => {
   try {
-    const token = await AsyncStorage.getItem(ACCESS_TOKEN);
-    return token;
+    const token = await AsyncStorage.getItem(name);
+    return token ? JSON.parse(token) : {};
   } catch {
     console.log("Storage error");
     return null;
   }
 };
 
-export const setTokenFromStorage = async (value) => {
+export const setTokenFromStorage = async (token, value) => {
   try {
-    await AsyncStorage.setItem(ACCESS_TOKEN, value);
+    await AsyncStorage.setItem(token, value);
     return true;
   } catch {
     console.log("Storage error");
@@ -34,22 +36,30 @@ export const getFileFromUri = (uri) => {
   };
 };
 
-export const handleLogout = async (dispatch) => {
-  try {
-    await AsyncStorage.clear();
-    dispatch({
-      type: SET_IS_LOGIN,
-      data: false,
-    });
-    dispatch({
-      type: SET_ACCESS_TOKEN,
-      data: "",
-    });
-  } catch {
-    console.log("Storage error");
+export const handleLogout = async () => {
+  const response = await UserApi.logout();
+
+  if (response.message === STATUS_MESSAGE.SUCCESS) {
+    const { dispatch } = store;
+
+    try {
+      await AsyncStorage.clear();
+      dispatch({
+        type: SET_IS_LOGIN,
+        data: false,
+      });
+    } catch {
+      console.log("Storage error");
+    }
   }
 };
 
 export const addPrefixUrl = (url) => {
   return !isEmpty(url) ? `${URL_SERVER}${url}` : "";
+};
+
+export const checkExpiredToken = (expiry) => {
+  const today = new Date();
+  const timeExpiredToken = new Date(expiry);
+  return timeExpiredToken >= today;
 };
